Make adjustments based on review comments.
